YES IT WOKRS

+++ fvm/mesh.py
@@ -6,7 +6,7 @@ import numpy as np
 class Vertex2D:
     def __init__(self, x=0, y=0):
         self.id = 0
-        self.coords = np.asarray([x, y])
+        self.coords = np.asarray([x, y], dtype=np.float64)
 
     def compute(self):
         pass
@@ -29,9 +29,9 @@ class Edge2D:
         self.id = 0
         self.vertexes_id = np.asarray([vertexes_1_idx, vertexes_2_idx])
         self.nodes_id = []
-        self.normal = np.asarray([0, 0])
-        self.t = np.asarray([0, 0])
-        self.center_coords = np.asarray([0, 0])
+        self.normal = np.asarray([0, 0], dtype=np.float64)
+        self.t = np.asarray([0, 0], dtype=np.float64)
+        self.center_coords = np.asarray([0, 0], dtype=np.float64)
         self.area = 0
 
     def compute(self):
@@ -41,14 +41,14 @@ class Edge2D:
         self.normal = np.asarray([
             -(p2_coords[1] - p1_coords[1]),
             (p2_coords[0] - p1_coords[0]),
-        ])
+        ], dtype=np.float64)
         self.normal = self.normal / np.linalg.norm(self.normal)
         self.t = np.asarray([
             (p2_coords[0] - p1_coords[0]),
             (p2_coords[1] - p1_coords[1]),
         ])
         self.t = self.t / np.linalg.norm(self.t)
-        self.area = np.sqrt(np.sum((p2_coords - p1_coords) ** 2))
+        self.area = np.sqrt(np.sum((p2_coords - p1_coords) ** 2, dtype=np.float64), dtype=np.float64)
 
     def __repr__(self):
         return f"Edge2D(area={self.area}, p1=({self.mesh.vertexes[self.vertexes_id[0]].coords}), p2=({self.mesh.vertexes[self.vertexes_id[1]].coords}), is_bound={len(self.nodes_id) < 2})"
@@ -60,8 +60,8 @@ class Quadrangle2D:
         self.id = 0
         self.edges_id = np.asarray([e1,e2,e3,e4])
         self.vertexes_id = np.asarray([v1,v2,v3,v4])
-        self.center_coords = np.asarray([0, 0])
-        self.vectors_in_edges_directions = np.zeros((4, 2), dtype=np.float16)
+        self.center_coords = np.asarray([0, 0], dtype=np.float64)
+        self.vectors_in_edges_directions = np.zeros((4, 2), dtype=np.float64)
         self.normals = []
         self.volume = 0
 
@@ -84,7 +84,7 @@ class Quadrangle2D:
             direction_vector = np.asarray([
                 (edge.center_coords[0] - self.center_coords[0]),
                 (edge.center_coords[1] - self.center_coords[1]),
-            ])
+            ], dtype=np.float64)
             direction_vector = direction_vector / np.linalg.norm(direction_vector)
             self.vectors_in_edges_directions[i] = direction_vector
         self.volume = 0.5 * (
@@ -97,12 +97,12 @@ class Quadrangle2D:
             v0, v1 = self.mesh.vertexes[v0_id], self.mesh.vertexes[v1_id]
             dx = v1.x - v0.x
             dy = v1.y - v0.y
-            self.normals.append(np.asarray([dy, -dx]))
+            self.normals.append(np.asarray([dy, -dx], dtype=np.float64))
         v0_id, v1_id = self.vertexes_id[-1], self.vertexes_id[0]
         v0, v1 = self.mesh.vertexes[v0_id], self.mesh.vertexes[v1_id]
         dx = v1.x - v0.x
         dy = v1.y - v0.y
-        self.normals.append(np.asarray([dy, -dx]))
+        self.normals.append(np.asarray([dy, -dx], dtype=np.float64))
 
 
     def __repr__(self):
@@ -128,7 +128,7 @@ class Quadrangle2DMesh:
         self.vertexes: List[Vertex2D] = []
         self.edged: List[Edge2D] = []
         self.nodes: List[Quadrangle2D] = []
-        self.volumes: np.ndarray = np.zeros(self.num_nodes, dtype=np.float16)
+        self.volumes: np.ndarray = np.zeros(self.num_nodes, dtype=np.float64)
         self._init_internals()
 
     def coord_fo_idx(self, x, y):
@@ -138,8 +138,8 @@ class Quadrangle2DMesh:
         return idx // self.x, idx % self.x
 
     def get_meshgrid(self):
-        x, y = np.linspace(0, self.lx, self.x + 1), np.linspace(0, self.ly, self.y + 1)
-        yv, xv = np.meshgrid(x, y)
+        x, y = np.linspace(0, self.lx, self.x + 1, dtype=np.float64), np.linspace(0, self.ly, self.y + 1, dtype=np.float64)
+        yv, xv = np.meshgrid(x, y, dtype=np.float64)
         for node in self.nodes:
             node_y, node_x = self.idx_to_coord(node.id)
             xv[node_x, node_y] = self.vertexes[node.vertexes_id[0]].coords[0]
